node3/requirements.txt
Flask>=2.2.5
flask_sqlalchemy==3.1.1
openai==0.28
python-dotenv
